move pages into content dir
