+++ app/blog/page.tsx
@@ -1,0 +1,1 @@
+
