--- content/hello-world.mdx
@@ -1,10 +1,0 @@
----
-title: No Meetings, No Deadlines, No Full-Time Employees
-description: I started Gumroad in 2011. In 2015, we reached a peak of 23 full-time employees. In 2016, after failing to raise more money, I ended up back where I began.
----
-
-I started Gumroad in 2011. In 2015, we reached a peak of 23 full-time employees. In 2016, after failing to raise more money, I ended up back where I began: a one-person company.
-
-Today, when I'm asked how many people work at Gumroad, I respond with "ten or so." That's how I convert the number of people we have into what others expect...
-
-[Rest of your content...]
