+++ package.json
@@ -22,6 +22,7 @@
     "gray-matter": "^4.0.3",
     "lucide-react": "^0.453.0",
     "next": "^14.2.15",
+    "next-mdx-remote": "^4.4.1",
     "postcss": "^8.4.41",
     "prettier": "^3.3.3",
     "prop-types": "^15.8.1",
@@ -29,8 +30,7 @@
     "react-dom": "^18.3.1",
     "react-ga": "^3.0.0",
     "tailwind-merge": "^2.5.4",
-    "tailwindcss-animate": "^1.0.7",
-    "next-mdx-remote": "^4.4.1"
+    "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
     "@types/node": "22.7.5",
